Isolate plugin dependency tests with worker harness
